Separate other libraries into another component to prevent unnecessary setup calls when not rendering them.
